move all the timepanel code
