+++ crates/re_time_panel/Cargo.toml
@@ -16,8 +16,18 @@ include = ["../../LICENSE-APACHE", "../../LICENSE-MIT", "**/*.rs", "Cargo.toml"]
 all-features = true
 
 [dependencies]
+re_arrow_store.workspace = true
+re_data_store.workspace = true
+re_data_ui.workspace = true
+re_log_types.workspace = true
 re_ui.workspace = true
 re_viewer_context.workspace = true
+
+egui.workspace = true
+itertools.workspace = true
+time.workspace = true
+vec1 = "1.8"
+serde = "1.0"
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 puffin.workspace = true
