+++ crates/re_time_panel/src/data_density_graph.rs
@@ -12,8 +12,6 @@ use re_data_ui::DataUi;
 use re_log_types::{TimeInt, TimeRange, TimeReal};
 use re_viewer_context::{Item, UiVerbosity, ViewerContext};
 
-use crate::ui::Blueprint;
-
 use super::time_ranges_ui::TimeRangesUi;
 
 // ----------------------------------------------------------------------------
@@ -365,7 +363,6 @@ fn smooth(density: &[f32]) -> Vec<f32> {
 pub fn data_density_graph_ui(
     data_dentity_graph_painter: &mut DataDensityGraphPainter,
     ctx: &mut ViewerContext<'_>,
-    blueprint: &mut Blueprint,
     time_area_response: &egui::Response,
     time_area_painter: &egui::Painter,
     ui: &mut egui::Ui,
@@ -489,7 +486,6 @@ pub fn data_density_graph_ui(
         } else if !ui.ctx().memory(|mem| mem.is_anything_being_dragged()) {
             show_row_ids_tooltip(
                 ctx,
-                blueprint,
                 ui.ctx(),
                 &item,
                 hovered_time_range,
@@ -519,7 +515,6 @@ fn make_brighter(color: Color32) -> Color32 {
 
 fn show_row_ids_tooltip(
     ctx: &mut ViewerContext<'_>,
-    blueprint: &mut Blueprint,
     egui_ctx: &egui::Context,
     item: &Item,
     time_range: TimeRange,
@@ -537,7 +532,8 @@ fn show_row_ids_tooltip(
         }
 
         ui.add_space(8.0);
-        crate::ui::selection_panel::what_is_selected_ui(ui, ctx, blueprint, item);
+        // TODO:
+        //crate::ui::selection_panel::what_is_selected_ui(ui, ctx, viewport, item);
         ui.add_space(8.0);
 
         let timeline = *ctx.rec_cfg.time_ctrl.timeline();
