+++ crates/re_time_panel/src/format_time.rs
