+++ crates/re_time_panel/src/paint_ticks.rs
@@ -4,7 +4,7 @@ use egui::{lerp, pos2, remap_clamp, Align2, Color32, Rect, Rgba, Shape, Stroke};
 
 use re_log_types::{Time, TimeRangeF, TimeReal, TimeType};
 
-use crate::misc::format_time::next_grid_tick_magnitude_ns;
+use crate::format_time::{format_time_compact, next_grid_tick_magnitude_ns};
 
 use super::time_ranges_ui::TimeRangesUi;
 
@@ -69,7 +69,7 @@ fn paint_time_range_ticks(
                 &ui.clip_rect(),
                 time_range, // ns
                 next_grid_tick_magnitude_ns,
-                |ns| crate::misc::format_time::format_time_compact(Time::from_ns_since_epoch(ns)),
+                |ns| format_time_compact(Time::from_ns_since_epoch(ns)),
             )
         }
         TimeType::Sequence => {
